refactor position logic
